refactor: separate packages for dtos, exceptions in service layer
