+++ src/main/java/techcourse/myblog/service/dto/UserRequestDto.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service;
+package techcourse.myblog.service.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
