+++ src/main/java/techcourse/myblog/service/exception/ArticleNotFoundException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service;
+package techcourse.myblog.service.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
