+++ src/main/java/techcourse/myblog/service/exception/AuthenticationFailException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service;
+package techcourse.myblog.service.exception;
 
 public class AuthenticationFailException extends RuntimeException {
     public AuthenticationFailException() {
